requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
WTForms
gunicorn
psycopg2-binary
python-dotenv
